+++ src/datasources/asset/components/editors/calibration-forecast/CalibrationForecastEditor.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 
-import { AssetCalibrationQuery } from '../../../../asset-calibration/types';
-import { AssetQuery } from '../../../types';
 import { CalibrationForecastDataSource } from './CalibrationForecastDataSource';
+import { AssetQuery } from '../../../types/types';
+import { CalibrationForecastQuery } from '../../../types/CalibrationForecastQuery.types';
 
 type Props = {
-  query: AssetCalibrationQuery;
+  query: CalibrationForecastQuery;
   handleQueryChange: (value: AssetQuery, runQuery: boolean) => void;
   datasource: CalibrationForecastDataSource;
 };
 
 export function CalibrationForecastEditor({ query, handleQueryChange, datasource }: Props) {
-  query = datasource.prepareQuery(query) as AssetCalibrationQuery;
+  query = datasource.prepareQuery(query) as CalibrationForecastQuery;
 
   return <div style={{ position: 'relative' }}>Calibration Forecast</div>;
 }
