+++ src/datasources/asset/module.ts
@@ -1,8 +1,8 @@
 import { DataSourcePlugin } from '@grafana/data';
 import { AssetDataSource } from './AssetDataSource';
-import { AssetCoordonatorQueryEditor } from './components/AssetCoordonatorQueryEditor';
+import { AssetQueryEditor } from './components/AssetQueryEditor';
 import { HttpConfigEditor } from 'core/components/HttpConfigEditor';
 
 export const plugin = new DataSourcePlugin(AssetDataSource)
   .setConfigEditor(HttpConfigEditor)
-  .setQueryEditor(AssetCoordonatorQueryEditor);
+  .setQueryEditor(AssetQueryEditor);
