--- src/datasources/asset/types.ts
@@ -1,65 +1,0 @@
-import { FieldDTO } from '@grafana/data'
-import { DataQuery } from '@grafana/schema'
-
-export interface ListAssetsQuery extends DataQuery {
-  workspace: string,
-  minionIds: string[]
-}
-
-export interface CalibrationForecastQuery extends DataQuery {
-}
-
-export interface AssetSummaryQuery extends DataQuery {
-}
-
-export enum AssetFilterProperties {
-  AssetIdentifier = 'AssetIdentifier',
-  SerialNumber = 'SerialNumber',
-  ModelName = 'ModelName',
-  VendorName = 'VendorName',
-  VendorNumber = 'VendorNumber',
-  AssetName = 'AssetName',
-  FirmwareVersion = 'FirmwareVersion',
-  HardwareVersion = 'HardwareVersion',
-  BusType = 'BusType',
-  IsNIAsset = 'IsNIAsset',
-  Keywords = 'Keywords',
-  Properties = 'Properties',
-  LocationMinionId = 'Location.MinionId',
-  LocationSlotNumber = 'Location.SlotNumber',
-  LocationAssetStateSystemConnection = 'Location.AssetState.SystemConnection',
-  LocationAssetStateAssetPresence = 'Location.AssetState.AssetPresence',
-  SupportsSelfCalibration = 'SupportsSelfCalibration',
-  SelfCalibrationCalibrationDate = 'SelfCalibration.CalibrationDate',
-  SupportsExternalCalibration = 'SupportsExternalCalibration',
-  CustomCalibrationInterval = 'CustomCalibrationInterval',
-  CalibrationStatus = 'CalibrationStatus',
-  ExternalCalibrationCalibrationDate = 'ExternalCalibration.CalibrationDate',
-  ExternalCalibrationNextRecommendedDate = 'ExternalCalibration.NextRecommendedDate',
-  ExternalCalibrationRecommendedInterval = 'ExternalCalibration.RecommendedInterval',
-  ExternalCalibrationComments = 'ExternalCalibration.Comments',
-  ExternalCalibrationIsLimited = 'ExternalCalibration.IsLimited',
-  ExternalCalibrationOperatorDisplayName = 'ExternalCalibration.Operator.DisplayName',
-  IsSystemController = 'IsSystemController'
-}
-
-export interface CalibrationForecastResponse {
-  calibrationForecast: CalibrationForecastModel
-}
-
-export interface CalibrationForecastModel {
-  columns: FieldDTO[],
-}
-
-export enum EntityType {
-  Asset = "Asset",
-  System = "System"
-}
-
-export enum AssetQueryType {
-  ListAssets = "List Assets",
-  CalibrationForecast = "Calibration Forecast",
-  AssetSummary = "Asset Summary"
-}
-
-export type AssetQuery = ListAssetsQuery | CalibrationForecastQuery | AssetSummaryQuery;
