+++ src/datasources/asset/types/AssetSummaryQuery.types.ts
@@ -1,0 +1,4 @@
+import { DataQuery } from '@grafana/schema'
+
+export interface AssetSummaryQuery extends DataQuery {
+}
